Use real framework code for creating system services.
Remove all api specific service creation code from ShadowContextImpl and switch it back from a velocity template to a to a pure Java class.

Have ShadowSystemManager return a dynamic proxy of the correct service type returning default values for SystemManager.getService() this will prevent a lot of NPE's when the underlying service is dereferenced in shadows and will safe stubbing stuff out unnecessarily.